Relocate generate date method to controllers. Server will now be responsible for generating the date. Update tests to import generate date from the correct file.
