+++ src/server/controller/plugin.ts
@@ -91,4 +91,14 @@ export default class Controller {
       throw err;
     }
   }
+
+  public generateDate(): string {
+    const today: Date = new Date();
+    const dd: String = String(today.getDate()).padStart(2, "0");
+    const mm: String = String(today.getMonth() + 1).padStart(2, "0");
+    const yyyy: String = String(today.getFullYear());
+    const date: string = `${mm}-${dd}-${yyyy}`;
+
+    return date;
+  }
 }
