+++ test/server/gateways/postgres/plugin.test.ts
@@ -1,12 +1,14 @@
 import pg from "pg";
 import { Trade } from "../../../../src/server/types";
 import { PostgresGateway } from "./../../../../src/server/gateways";
+import Controller from "../../../../src/server/controller/index";
 
 describe("Postgres Gateway Tests", function () {
   jest.mock("pg");
 
+  const controller = new Controller();
   const db = new PostgresGateway();
-  const now: string = db.generateDate();
+  const now: string = controller.generateDate();
   const trade: Trade = {
     id: "abc",
     ticker: "ABC",
